MOB-1160 - moved the select ACL checks to the nodes that are doing the  lookups so that ACL access is local. Also extended HttpClient to use new select reader and fixed some tests.
